# docs/openapi-chatgpt-actions.yaml
# OpenAPI 3.1 schema for ChatGPT Custom GPT Actions
# Import this when creating a Custom GPT at platform.openai.com
# ChatGPT requires openapi: 3.1.0 or 3.1.1
openapi: 3.1.0
info:
  title: USO Product Discovery
  description: Discover products from natural language. For ChatGPT Custom GPT.
  version: 1.0.0
servers:
  - url: https://uso-orchestrator.onrender.com
    description: Render staging
paths:
  /api/v1/chat:
    post:
      operationId: chat
      summary: Discover products from natural language
      description: |
        Send user message (e.g. "find flowers") and get intent + products.
        Returns a human-readable summary and product details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: User message (e.g. "find flowers")
                user_id:
                  type: string
                  nullable: true
                  description: Optional user ID
                limit:
                  type: integer
                  default: 20
                  minimum: 1
                  maximum: 100
                  description: Max products to return
                thread_id:
                  type: string
                  nullable: true
                  description: Chat thread ID for status updates (optional)
                platform:
                  type: string
                  enum: [chatgpt, gemini]
                  nullable: true
                  description: Platform for webhook push (optional)
      responses:
        "200":
          description: Intent and products with human-readable summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Human-readable summary for chat display
                  data:
                    type: object
                    properties:
                      intent:
                        type: object
                        properties:
                          intent_type: { type: string }
                          search_query: { type: string }
                      products:
                        type: object
                        properties:
                          products: { type: array }
                          count: { type: integer }
                  adaptive_card:
                    type: object
                    description: Adaptive Card for rich UI
                  machine_readable:
                    type: object
                    description: JSON-LD for LLM consumption
  /api/v1/products/{product_id}:
    get:
      operationId: getProductDetails
      summary: Get product details by ID
      description: |
        Fetch full product details for View Details action.
        Use when user clicks "View Details" on a product card.
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: string
          description: Product UUID
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      description: { type: string }
                      price: { type: number }
                      currency: { type: string }
        "404":
          description: Product not found
  /api/v1/bundles/{bundle_id}:
    get:
      operationId: getBundleDetails
      summary: Get bundle details by ID
      description: |
        Fetch bundle with items for View Bundle action.
        Use when user wants to see their bundle/cart contents.
      parameters:
        - name: bundle_id
          in: path
          required: true
          schema:
            type: string
          description: Bundle UUID
      responses:
        "200":
          description: Bundle with items
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Human-readable bundle summary
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      total_price: { type: number }
                      currency: { type: string }
                      item_count: { type: integer }
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: string }
                            name: { type: string }
                            price: { type: number }
                  adaptive_card:
                    type: object
                    description: Bundle Adaptive Card
        "404":
          description: Bundle not found
  /api/v1/bundle/add:
    post:
      operationId: addToBundle
      summary: Add product to bundle
      description: |
        Add a product to the user's bundle (cart).
        Use when user clicks "Add to Bundle" on a product card.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id]
              properties:
                product_id:
                  type: string
                  description: Product UUID from product card
                user_id:
                  type: string
                  nullable: true
                  description: Optional user ID
                bundle_id:
                  type: string
                  nullable: true
                  description: Existing bundle ID to add to (creates new if omitted)
      responses:
        "200":
          description: Product added to bundle
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Human-readable confirmation
                  data:
                    type: object
                    properties:
                      bundle_id: { type: string }
                      product_added: { type: string }
                      total_price: { type: number }
                      currency: { type: string }
        "400":
          description: Failed to add (product not found or error)
  /api/v1/bundle/remove:
    post:
      operationId: removeFromBundle
      summary: Remove item from bundle
      description: |
        Remove an item from the user's bundle (cart).
        Use when user clicks "Remove" on an item in the bundle card.
        item_id is the bundle leg id from the bundle's items (each item has an id field).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item_id]
              properties:
                item_id:
                  type: string
                  description: Bundle leg id (item id from bundle card, e.g. from getBundleDetails response data.items[].id)
      responses:
        "200":
          description: Item removed from bundle
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Human-readable confirmation
                  data:
                    type: object
                    properties:
                      bundle_id: { type: string }
                      item_removed_id: { type: string }
                      total_price: { type: number }
                      currency: { type: string }
        "404":
          description: Item or bundle not found
  /api/v1/checkout:
    post:
      operationId: proceedToCheckout
      summary: Proceed to checkout with bundle
      description: |
        Initiate checkout for a bundle (cart). Creates order, returns order_id.
        Use when user clicks "Proceed to Checkout" on the bundle card.
        Returns order summary and checkout card. Use order_id with createPaymentIntent to pay.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bundle_id]
              properties:
                bundle_id:
                  type: string
                  description: Bundle UUID from bundle card
      responses:
        "200":
          description: Checkout initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    description: Human-readable confirmation
                  order_id:
                    type: string
                    description: Order UUID for payment (use with createPaymentIntent)
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      bundle_id: { type: string }
                      total: { type: number }
                      currency: { type: string }
                      status: { type: string }
                      line_items: { type: array }
                  adaptive_card:
                    type: object
                    description: Checkout Adaptive Card
        "400":
          description: Bundle is empty
        "404":
          description: Bundle not found
  /api/v1/payment/create:
    post:
      operationId: createPaymentIntent
      summary: Create payment intent for order
      description: |
        Create Stripe PaymentIntent for an order. Use order_id from proceedToCheckout response.
        Call when user wants to pay (e.g. "pay now", "complete payment").
        Returns client_secret for payment confirmation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id]
              properties:
                order_id:
                  type: string
                  description: Order UUID from checkout response
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id: { type: string }
                  client_secret: { type: string }
                  payment_intent_id: { type: string }
                  amount: { type: number }
                  currency: { type: string }
        "404":
          description: Order not found
        "400":
          description: Order already paid or amount too small
  /api/v1/payment/confirm:
    post:
      operationId: confirmPayment
      summary: Confirm payment for order (demo mode)
      description: |
        Mark order as paid. Use when DEMO_PAYMENT=true on the server (for testing).
        Call after createPaymentIntent when user says "yes" or "complete payment".
        In production, payment is completed via Stripe checkout page + webhook.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id]
              properties:
                order_id:
                  type: string
                  description: Order UUID from checkout
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id: { type: string }
                  status: { type: string }
                  message: { type: string }
        "403":
          description: Demo mode disabled (DEMO_PAYMENT not set)
        "404":
          description: Order not found
        "400":
          description: Order already paid
  /api/v1/change-request:
    post:
      operationId: requestChange
      summary: Request change from partner
      description: |
        Create change request and notify partner (e.g. "pink instead of red").
        Use when user requests a product change on an order.
        Requires order_id, order_leg_id, partner_id, original_item, requested_change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, order_leg_id, partner_id, original_item, requested_change]
              properties:
                order_id: { type: string }
                order_leg_id: { type: string }
                partner_id: { type: string }
                original_item:
                  type: object
                  properties:
                    product_id: { type: string }
                    name: { type: string }
                    price: { type: number }
                requested_change:
                  type: object
                  properties:
                    description: { type: string }
                respond_by: { type: string }
      responses:
        "200":
          description: Change request sent to partner
          content:
            application/json:
              schema:
                type: object
                properties:
                  negotiation_id: { type: string }
                  status: { type: string }
                  message: { type: string }
        "400":
          description: Partner has no webhook configured